fix: pass all tests

+++ scr/AccountManager.cpp
@@ -129,5 +129,20 @@ bool AccountManager::deleteUser(const std::string &userID) {
     if (currentPrivilege() != 7) {
         return false;
     }
+
+    // 1. 用户必须存在
+    UserRecord tmp;
+    if (!db.getUser(userID, tmp)) {
+        return false;
+    }
+
+    // 2. 用户不能处于登录状态（在 loginStack 中）
+    for (auto &sess : loginStack) {
+        if (sess.user.userID == userID) {
+            return false;
+        }
+    }
+
+    // 3. 删除
     return db.deleteUser(userID);
 }
